+++ Practica1/engine/src/main/java/com/gavilanvillar/engine/Rect.java
@@ -13,7 +13,7 @@ public class Rect {
      * @param _top Posición superior del rectángulo
      * @param _bottom Posición inferior del rectángulo
      */
-    public Rect(int _left, int _right, int _top, int _bottom) {
+    public Rect(float  _left, float  _right, float  _top, float  _bottom) {
         this._left = _left;
         this._right = _right;
         this._width = this._right - this._left;
@@ -26,12 +26,12 @@ public class Rect {
     //- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
     //        Atributos públicos (de Rect)
     //- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-    public int _left;
-    public int _right;
-    public int _width;
+    public float _left;
+    public float  _right;
+    public float  _width;
 
-    public int _top;
-    public int _bottom;
-    public int _height;
+    public float  _top;
+    public float  _bottom;
+    public float  _height;
 
 } // class Rect
